fix: move selector helper functions into SelectHelper class
